Remove duplicate ChangeViewModel tests and consolidate
